backend/requirements.txt
django-colorfield==0.7.2
django-cors-headers==3.13.0
Pillow==8.3.1
Django==3.2.13
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
webcolors==1.11.1
gunicorn==20.0.4
psycopg2-binary==2.9.3
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
reportlab==3.6